modified general interface
